Created a new controller for the messages
